+++ relayer-cli/src/commands/query.rs
@@ -6,6 +6,7 @@ use abscissa_core::{Command, Options, Runnable};
 mod channel;
 mod channels;
 mod client;
+mod clients;
 mod connection;
 mod connections;
 mod packet;
@@ -18,7 +19,7 @@ pub enum QueryCmd {
     Client(QueryClientCmds),
 
     #[options(help = "query clients")]
-    Clients(client::QueryAllClientsCmd),
+    Clients(clients::QueryAllClientsCmd),
 
     /// The `query connection` subcommand
     #[options(help = "query information about connection(s)")]
