+++ relayer-cli/src/commands/query/client.rs
@@ -11,7 +11,6 @@ use ibc::ics24_host::error::ValidationError;
 use ibc::ics24_host::identifier::ChainId;
 use ibc::ics24_host::identifier::ClientId;
 use ibc::ics24_host::Path::{ClientConnections, ClientConsensusState, ClientState};
-use ibc_proto::ibc::core::client::v1::QueryClientStatesRequest;
 use relayer::chain::Chain;
 use relayer::chain::CosmosSDKChain;
 use relayer::config::{ChainConfig, Config};
@@ -19,58 +18,6 @@ use relayer::config::{ChainConfig, Config};
 use crate::conclude::Output;
 use crate::error::{Error, Kind};
 use crate::prelude::*;
-
-/// Query clients command
-#[derive(Clone, Command, Debug, Options)]
-pub struct QueryAllClientsCmd {
-    #[options(free, help = "identifier of the chain to query")]
-    chain_id: ChainId,
-}
-
-impl QueryAllClientsCmd {
-    fn validate_options(&self, config: &Config) -> Result<ChainConfig, String> {
-        let chain_id = self.chain_id.clone();
-
-        let chain_config = config
-            .find_chain(&chain_id)
-            .ok_or_else(|| "missing chain configuration for the given chain id".to_string())?;
-
-        Ok(chain_config.clone())
-    }
-}
-
-/// Command for querying a client's state.
-/// To run with proof:
-/// rrly -c cfg.toml query client all ibc-1 --height 3
-///
-/// Run without proof:
-/// rrly -c cfg.toml query client all ibc-1 --height 3 -p false
-impl Runnable for QueryAllClientsCmd {
-    fn run(&self) {
-        let config = app_config();
-
-        let chain_config = match self.validate_options(&config) {
-            Err(err) => {
-                return Output::error(err).exit();
-            }
-            Ok(result) => result,
-        };
-
-        let rt = Arc::new(TokioRuntime::new().unwrap());
-        let chain = CosmosSDKChain::bootstrap(chain_config, rt).unwrap();
-
-        let req = QueryClientStatesRequest { pagination: None };
-
-        let res: Result<_, Error> = chain
-            .query_clients(req)
-            .map_err(|e| Kind::Query.context(e).into());
-
-        match res {
-            Ok(cids) => Output::success(cids).exit(),
-            Err(e) => Output::error(format!("{}", e)).exit(),
-        }
-    }
-}
 
 /// Query client state command
 #[derive(Clone, Command, Debug, Options)]
