+++ relayer/src/chain/mock.rs
@@ -238,7 +238,7 @@ impl Chain for MockChain {
         unimplemented!()
     }
 
-    fn query_clients(&self, _request: QueryClientStatesRequest) -> Result<Vec<ClientId>, Error> { 
+    fn query_clients(&self, _request: QueryClientStatesRequest) -> Result<Vec<ClientId>, Error> {
         unimplemented!()
     }
